New Repo and Enums folder and Repo classes
I got ahead of myself and made changes without committing prior changes,
so this commit is rather beefy.

A new method was added to the IRepo Interface. The method returns the
last address, which is the current address of the user.

The Program class is also changed to facilitate testing.

These changes should have been committed prior to the most recent
changes.

The remainder of the changes create new folders called Enums and Repos.
A new FieldType enum is created, which is used when updating an address.
The enum lists all of the fields.

Another repo class and associated interface is created as well. These
files are created in the new Repos folder. In addition, the existing
repo class and its associated interface is moved to the new Repos
folder.

The former DataRepo class has been moved to the new BaseRepo class. The
DataRepo class has been repurposed for an additional layer of code. The
BaseRepo class provides the actual interface to Entity Framework and the
database. The new layer, provided in the DataRepo class, provides a
layer which add logic necessary to process a request, yet is in addition
to the low-level access to the database.

The Program class is further impacted by moving code from that class
into the new DataRepo class. The moves better align the project with
the MVC pattern and separation of concerns principle.

Signed-off-by: Grauldon <17791090+Grauldon@users.noreply.github.com>

  Changes to be committed:
 	modified:   AddressHistory/Program.cs
 	new file:   AddressHistoryDAL/Enums/FieldType.cs
 	renamed:    AddressHistoryDAL/DataRepo.cs -> AddressHistoryDAL/Repos/BaseRepo.cs
 	new file:   AddressHistoryDAL/Repos/DataRepo.cs
 	new file:   AddressHistoryDAL/Repos/IDataRepo.cs
 	renamed:    AddressHistoryDAL/IRepo.cs -> AddressHistoryDAL/Repos/IRepo.cs