Improve the InsertAddress method
The InsertMethod is improved to include logic to ensure the current
address, which is to be terminated, is not effective with the current
date since the new address will have a start date of the current date.

The TerminateAddress function is removed. The code is merged with the
InsertAddress method and modified where necessary to fit in the new
logic flow.

The decision to merge is based on the InsertAddress method now requiring
the last address. Instead of two calls against the table, the functions
are merged so one call may be made.

The code is also changed so that the StartDate and EndDate properties are
populated by the InsertAddress method instead of the values being passed
in (at least, any values passed in will be overwritten). The values of
these properties are set by design and do not change based on external
factors. Having them set in this method ensures consistency and data
integrity.

Signed-off-by: Grauldon <17791090+Grauldon@users.noreply.github.com>

  Changes to be committed:
 	modified:   Program.cs
 	modified:   ../AddressHistoryDAL/Repos/DataRepo.cs